refactor ntl utils
